@@ -1,61 +1,74 @@
 import emitter from '../emitter';
-import { canUseDOM } from 'fbjs/lib/ExecutionEnvironment';
 
-let playSubscription, winSubscription;
+let enabled = false;
+let playSubscription = null;
+let winSubscription = null;
 
 export default {
   enable() {
-    if (canUseDOM) {
-      if (typeof mixpanel === 'undefined') {
+    if (enabled) {
+      return;
+    }
+
+    // Check if we're in a browser environment
+    const isBrowser = typeof window !== 'undefined';
+
+    if (isBrowser) {
+      if (!window.mixpanel) {
         const error = new Error(
           "React A/B Test Mixpanel Helper: 'mixpanel' global is not defined."
         );
         error.type = 'PUSHTELL_HELPER_MISSING_GLOBAL';
         throw error;
       }
-      playSubscription = emitter.addPlayListener(function (
-        experimentName,
-        variantName
-      ) {
-        mixpanel.track(
-          'Experiment Play',
-          {
-            Experiment: experimentName,
-            Variant: variantName,
-          },
-          function () {
-            emitter.emit('mixpanel-play', experimentName, variantName);
-          }
-        );
-      });
-      winSubscription = emitter.addWinListener(function (
-        experimentName,
-        variantName
-      ) {
-        mixpanel.track(
-          'Experiment Win',
-          {
-            Experiment: experimentName,
-            Variant: variantName,
-          },
-          function () {
-            emitter.emit('mixpanel-win', experimentName, variantName);
-          }
-        );
-      });
+    } else {
+      // If not in browser, just enable without functionality
+      enabled = true;
+      return;
     }
+
+    playSubscription = emitter.addPlayListener(function (
+      experimentName,
+      variantName
+    ) {
+      window.mixpanel.track('Experiment Viewed', {
+        experimentName: experimentName,
+        variationName: variantName,
+      });
+    });
+
+    winSubscription = emitter.addWinListener(function (
+      experimentName,
+      variantName
+    ) {
+      window.mixpanel.track('Experiment Won', {
+        experimentName: experimentName,
+        variationName: variantName,
+      });
+    });
+
+    enabled = true;
   },
+
   disable() {
-    if (canUseDOM) {
-      if (!playSubscription || !winSubscription) {
-        const error = new Error(
-          'React A/B Test Mixpanel Helper: Helper was not enabled.'
-        );
-        error.type = 'PUSHTELL_HELPER_INVALID_DISABLE';
-        throw error;
-      }
-      playSubscription.remove();
-      winSubscription.remove();
+    if (!enabled) {
+      const error = new Error(
+        'React A/B Test Mixpanel Helper: Helper was disabled without being enabled first.'
+      );
+      error.type = 'PUSHTELL_HELPER_INVALID_DISABLE';
+      throw error;
     }
+
+    if (playSubscription) {
+      playSubscription.remove();
+      playSubscription = null;
+    }
+
+    if (winSubscription) {
+      winSubscription.remove();
+      winSubscription = null;
+    }
+
+    enabled = false;
   },
 };
