@@ -1,12 +1,12 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { v4 as UUID } from 'uuid';
-import { mount } from 'enzyme';
-import { act } from 'react-dom/test-utils';
+import { renderWithWrapper } from '../test-utils';
+import { act } from '@testing-library/react';
 
-import CoreExperiment from '../../src/CoreExperiment.jsx';
-import Experiment from '../../src/Experiment.jsx';
-import Variant from '../../src/Variant.jsx';
-import emitter from '../../src/emitter.jsx';
+import CoreExperiment from '../../src/CoreExperiment';
+import Experiment from '../../src/Experiment';
+import Variant from '../../src/Variant';
+import emitter from '../../src/emitter';
 
 describe('Emitter', () => {
   afterEach(() => {
@@ -15,21 +15,22 @@ describe('Emitter', () => {
 
   it('should throw an error when passed an invalid name argument.', () => {
     expect(() => {
+      // @ts-expect-error - intentionally passing wrong type for test
       emitter.emitWin(1);
-    }).toThrowError(/type \'string\'/);
+    }).toThrow(/type \'string\'/);
   });
 
   it('should emit when a variant is played.', () => {
     const experimentName = UUID();
     // TODO: use spies?
     let playedVariantName = null;
-    let playCallback = (experimentName, variantName) => {
+    const playCallback = (experimentName, variantName) => {
       playedVariantName = variantName;
     };
 
     let experimentNameGlobal = null;
     let playedVariantNameGlobal = null;
-    let playCallbackGlobal = (experimentName, variantName) => {
+    const playCallbackGlobal = (experimentName, variantName) => {
       experimentNameGlobal = experimentName;
       playedVariantNameGlobal = variantName;
     };
@@ -40,7 +41,7 @@ describe('Emitter', () => {
     );
     const playSubscriptionGlobal = emitter.addPlayListener(playCallbackGlobal);
 
-    const wrapper = mount(
+    renderWithWrapper(
       <CoreExperiment name={experimentName} defaultVariantName="A">
         <Variant name="A">
           <div id="variant-a" />
@@ -61,19 +62,19 @@ describe('Emitter', () => {
   it('should emit when a variant wins.', () => {
     const experimentName = UUID();
     let winningVariantName = null;
-    let winCallback = function (experimentName, variantName) {
+    const winCallback = (experimentName, variantName) => {
       winningVariantName = variantName;
     };
     let experimentNameGlobal = null;
     let winningVariantNameGlobal = null;
-    let winCallbackGlobal = function (experimentName, variantName) {
+    const winCallbackGlobal = (experimentName, variantName) => {
       experimentNameGlobal = experimentName;
       winningVariantNameGlobal = variantName;
     };
     const winSubscription = emitter.addWinListener(experimentName, winCallback);
     const winSubscriptionGlobal = emitter.addWinListener(winCallbackGlobal);
 
-    const wrapper = mount(
+    renderWithWrapper(
       <CoreExperiment name={experimentName} defaultVariantName="A">
         <Variant name="A">
           <div id="variant-a" />
@@ -95,35 +96,29 @@ describe('Emitter', () => {
   });
 
   it('should emit when a variant is clicked.', () => {
-    let experimentName = UUID();
+    const experimentName = UUID();
 
     let winningVariantName = null;
-    let winCallback = function (experimentName, variantName) {
+    const winCallback = (experimentName, variantName) => {
       winningVariantName = variantName;
     };
     let experimentNameGlobal = null;
     let winningVariantNameGlobal = null;
-    let winCallbackGlobal = function (experimentName, variantName) {
+    const winCallbackGlobal = (experimentName, variantName) => {
       experimentNameGlobal = experimentName;
       winningVariantNameGlobal = variantName;
     };
     const winSubscription = emitter.addWinListener(experimentName, winCallback);
     const winSubscriptionGlobal = emitter.addWinListener(winCallbackGlobal);
 
-    class App extends Component {
-      onClickVariant = (e) => {
-        this.experimentRef.current.win();
+    class App extends React.Component {
+      onClickVariant = () => {
+        emitter.emitWin(experimentName);
       };
-
-      experimentRef = React.createRef();
 
       render() {
         return (
-          <Experiment
-            ref={this.experimentRef}
-            name={experimentName}
-            defaultVariantName="A"
-          >
+          <Experiment name={experimentName} defaultVariantName="A">
             <Variant name="A">
               <a id="variant-a" href="#A" onClick={this.onClickVariant}>
                 A
@@ -139,8 +134,14 @@ describe('Emitter', () => {
       }
     }
 
-    const wrapper = mount(<App />);
-    wrapper.find('#variant-a').simulate('click');
+    let wrapper;
+    act(() => {
+      wrapper = renderWithWrapper(<App />);
+    });
+
+    act(() => {
+      wrapper.find('#variant-a').click();
+    });
 
     expect(winningVariantName).toBe('A');
     expect(experimentNameGlobal).toBe(experimentName);
@@ -170,14 +171,8 @@ describe('Emitter', () => {
       activeVariantCallbackGlobal
     );
 
-    const experimentRef = React.createRef();
-
-    const wrapper = mount(
-      <Experiment
-        ref={experimentRef}
-        name={experimentName}
-        defaultVariantName="A"
-      >
+    renderWithWrapper(
+      <Experiment name={experimentName} defaultVariantName="A">
         <Variant name="A">
           <a id="variant-a" href="#A">
             A
@@ -201,14 +196,9 @@ describe('Emitter', () => {
 
   it('should get the experiment value.', () => {
     const experimentName = UUID();
-    const experimentRef = React.createRef();
 
-    const wrapper = mount(
-      <Experiment
-        ref={experimentRef}
-        name={experimentName}
-        defaultVariantName="A"
-      >
+    renderWithWrapper(
+      <Experiment name={experimentName} defaultVariantName="A">
         <Variant name="A">
           <a id="variant-a" href="#A">
             A
@@ -228,31 +218,34 @@ describe('Emitter', () => {
   it('should update the rendered component.', () => {
     const experimentName = UUID();
 
-    const wrapper = mount(
-      <CoreExperiment name={experimentName} defaultVariantName="A">
-        <Variant name="A">
-          <div id="variant-a" />
-        </Variant>
-        <Variant name="B">
-          <div id="variant-b" />
-        </Variant>
-      </CoreExperiment>
-    );
+    let wrapper;
+    act(() => {
+      wrapper = renderWithWrapper(
+        <CoreExperiment name={experimentName} defaultVariantName="A">
+          <Variant name="A">
+            <div id="variant-a" />
+          </Variant>
+          <Variant name="B">
+            <div id="variant-b" />
+          </Variant>
+        </CoreExperiment>
+      );
+    });
 
-    expect(wrapper.find('#variant-a').exists());
-    expect(!wrapper.find('#variant-b').exists());
+    expect(wrapper.find('#variant-a').exists()).toBeTruthy();
+    expect(wrapper.find('#variant-b').exists()).toBeFalsy();
 
     act(() => {
       emitter.setActiveVariant(experimentName, 'B');
     });
 
-    expect(!wrapper.find('#variant-a').exists());
-    expect(wrapper.find('#variant-b').exists());
+    expect(wrapper.find('#variant-a').exists()).toBeFalsy();
+    expect(wrapper.find('#variant-b').exists()).toBeTruthy();
   });
 
   it('should report active components.', () => {
-    let experimentNameA = UUID();
-    let experimentNameB = UUID();
+    const experimentNameA = UUID();
+    const experimentNameB = UUID();
 
     const AppA = () => (
       <CoreExperiment name={experimentNameA} defaultVariantName="A">
@@ -283,8 +276,17 @@ describe('Emitter', () => {
       </div>
     );
 
-    let wrapper = mount(<AppA />);
-    expect(emitter.getActiveExperiments()).toEqual({
+    // Test each component separately
+    let wrapper;
+
+    // First, test AppA
+    emitter._reset(); // Ensure clean state
+    act(() => {
+      wrapper = renderWithWrapper(<AppA />);
+    });
+
+    const experimentsA = emitter.getActiveExperiments();
+    expect(experimentsA).toEqual({
       [experimentNameA]: {
         A: true,
         B: false,
@@ -292,8 +294,14 @@ describe('Emitter', () => {
     });
     wrapper.unmount();
 
-    wrapper = mount(<AppB />);
-    expect(emitter.getActiveExperiments()).toEqual({
+    // Then, test AppB
+    emitter._reset(); // Ensure clean state
+    act(() => {
+      wrapper = renderWithWrapper(<AppB />);
+    });
+
+    const experimentsB = emitter.getActiveExperiments();
+    expect(experimentsB).toEqual({
       [experimentNameB]: {
         C: true,
         D: false,
@@ -301,8 +309,14 @@ describe('Emitter', () => {
     });
     wrapper.unmount();
 
-    wrapper = mount(<AppCombined />);
-    expect(emitter.getActiveExperiments()).toEqual({
+    // Finally, test combined app
+    emitter._reset(); // Ensure clean state
+    act(() => {
+      wrapper = renderWithWrapper(<AppCombined />);
+    });
+
+    const experimentsCombined = emitter.getActiveExperiments();
+    expect(experimentsCombined).toEqual({
       [experimentNameA]: {
         A: true,
         B: false,
@@ -325,12 +339,40 @@ describe('Emitter', () => {
     const experimentName = UUID();
     emitter.defineVariants(experimentName, ['A', 'B']);
     emitter.setCustomDistributionAlgorithm((_experimentName, userIdentifier) =>
-      parseInt(userIdentifier) % 4 === 0 ? 'A' : 'B'
+      parseInt(userIdentifier || '0') % 4 === 0 ? 'A' : 'B'
     );
     const variants = [...Array(6).keys()].map((userIdentifier) =>
       emitter.calculateActiveVariant(experimentName, userIdentifier.toString())
     );
 
     expect(variants).toEqual(['A', 'B', 'B', 'B', 'A', 'B']);
+  });
+
+  it('should findDOMNode or the browser.', () => {
+    const experimentName = UUID();
+
+    let wrapper;
+    act(() => {
+      wrapper = renderWithWrapper(
+        <CoreExperiment name={experimentName} defaultVariantName="A">
+          <Variant name="A">
+            <div id="variant-a" />
+          </Variant>
+          <Variant name="B">
+            <div id="variant-b" />
+          </Variant>
+        </CoreExperiment>
+      );
+    });
+
+    expect(wrapper.find('#variant-a').exists()).toBeTruthy();
+    expect(wrapper.find('#variant-b').exists()).toBeFalsy();
+
+    act(() => {
+      emitter.setActiveVariant(experimentName, 'B');
+    });
+
+    expect(wrapper.find('#variant-a').exists()).toBeFalsy();
+    expect(wrapper.find('#variant-b').exists()).toBeTruthy();
   });
 });
